# *: do not use SortingWriter for compaction